Fix assertion link. Change PageObject example for Selectors
