+++ docs/articles/index.html
@@ -82,25 +82,8 @@ test('Emulate user actions and perform a verification', async t => {
                 <p>The selector API provides methods that can be combined together, thus providing you with a flexible functional-style selector mechanism.
                 <pre><code>const macOSInput = Selector('.column').find('label').withText('MacOS').child('input');</code></pre>
                 </p>
-                <p>You can also use the helpful and popular <a href="http://martinfowler.com/bliki/PageObject.html">Page Object</a> methodology:</p>
-                <pre><code>import { Selector } from 'testcafe';
-
-fixture `Getting Started`
-    .page `http://devexpress.github.io/testcafe/example`;
-
-const page = {
-    nameInput:             Selector('#developer-name'),
-    interfaceSelect:       Selector('#preferred-interface'),
-    interfaceSelectOption: Selector('#preferred-interface').child('option')
-};
-
-test('Use Page Object', async t => {
-    await t
-        .typeText(page.nameInput, 'John Smith')
-        .click(page.interfaceSelect)
-        .click(page.interfaceSelectOption.nth(2))
-        .expect(page.interfaceSelect.value).eql('Both');
-});</code></pre>
+                <p>Check our <a href="https://github.com/DevExpress/testcafe/blob/master/examples/basic">example</a> that demonstrates how to create tests using
+                declarative <a href="http://martinfowler.com/bliki/PageObject.html">Page Objects</a> built with TestCafe selectors.</p>
                 </div>
             </div>
         <div class="feature">
